requirements.txt
Keras==2.0.2
matplotlib==1.2.0
numpy==1.12.0
Pillow==4.0.0
scikit-learn==0.18.1
scipy==0.18.1
Theano==0.9.0
